# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/newPictures": {
            "get": {
                "tags": [
                    "moderator"
                ],
                "summary": "New pictures",
                "description": "Paintings for sale",
                "operationId": "f87e3f7cb750419ff558320b9f5d5622",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Портрет"
                                        },
                                        "user": {
                                            "type": "string",
                                            "example": "Соня"
                                        },
                                        "user_avatar": {
                                            "type": "string",
                                            "example": "/storage/img/avatar.png"
                                        },
                                        "genre": {
                                            "type": "string",
                                            "example": "Портрет"
                                        },
                                        "description": {
                                            "type": "text",
                                            "example": "Нарисовать портрет"
                                        },
                                        "price": {
                                            "type": "number",
                                            "format": "float",
                                            "example": "1000"
                                        },
                                        "discount": {
                                            "type": "integer",
                                            "example": "10"
                                        },
                                        "size": {
                                            "type": "string",
                                            "example": "200*200"
                                        },
                                        "writing_technique": {
                                            "type": "string",
                                            "example": "Масло"
                                        },
                                        "img": {
                                            "type": "string",
                                            "example": "/storage/img/cats.jpg"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2022-04-08T04:48:40.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/newPictures/{id}": {
            "put": {
                "tags": [
                    "moderator"
                ],
                "summary": "Allow",
                "description": "Allow sale",
                "operationId": "b6e9d11ca8208dd7134b0d4261559ba1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Picture Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Картина выставлена на продажу"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "moderator"
                ],
                "summary": "Reject",
                "description": "Refusal to sell",
                "operationId": "1ae964117bad71f92a7fe7bdb1fae505",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Picture Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "required, string, min:10, max:100"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "reason": "Недопустимое содержимое"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Отказ в продаже"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "scheme": "Bearer"
            }
        }
    }
}